+++ ChangeLog.md
@@ -6,9 +6,10 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 
 - Add command `/log copy <player>` or `/logCopy <player>`, to copy logger configurations from other players.
 - Modify command `/tick warp` with new options: `interrupt`, `status`, etc.
-- Add HUD logger `tickWarp` with options: `bar`, `value`.
-- Add logger field `NAME`.
+- Add field `NAME` for all the logHelpers.
 - Display the actual option when setting logger with default options.
-  - `Subscribed to <logger> \(<option>\)`
+    - `Subscribed to <logger> \(<option>\)`
 
 # Features
+
+- Add HUD logger `tickWarp` with options: `bar`, `value`.